==> Codes/old_codes/3c_cnn-Copy3_scale.ipynb <==
# https://machinelearningmastery.com/pytorch-tutorial-develop-deep-learning-models/
# https://colab.research.google.com/github/rpi-techfundamentals/website_spring_2020/blob/master/content/notebooks/20-deep-learning1/06-regression-bh-pytorch.ipynb#scrollTo=xD9PhAU7hoqT
#!pip install torchvision
import numpy as np
import pandas as pd
import pyreadr
import scipy
#Define the model 
import torch
import torch.nn as nn
import torch.nn.functional as F
from sklearn.utils import shuffle
from torch.autograd import Variable
high_var = ['rain_30', 'rain_60','rain_90', 't_max_30', 't_max_60', 't_max_90', 't_min_30', 't_min_60', 't_min_90', 'Y_prev',
 'Y_corn_lt_avg', 'day_sow', 'day_v5', 'lai_v5', 'whc', 'oc_20cm_v5', 'sw_dep_v5', 'n_0_60cm_v5', 'surfaceom_wt_v5', 'sand_40cm', 'clay_40cm']


# check pytorch version
import torch
print(torch.__version__)

class Net(torch.nn.Module):
    def __init__(self, cols):
        super(Net, self).__init__()
        self.hidden1 = torch.nn.Linear(cols, 50)   # hidden layer
        self.hidden2 = torch.nn.Linear(50, 10)   # hidden layer
        #self.hidden3 = torch.nn.Linear(100, 50)   # hidden layer
        self.predict = torch.nn.Linear(10, 1)   # output layer

    def forward(self, x):
        x = F.relu(self.hidden1(x))      # activation function for hidden layer
        x = F.relu(self.hidden2(x))      # activation function for hidden layer
        #x = F.relu(self.hidden3(x))      # activation function for hidden layer
        x = self.predict(x)             # linear output
        return x

def build_cnn(data, policy):
    #Define training hyperprameters.
    batch_size = 50
    num_epochs = 100
    learning_rate = 0.01
    
    y_train = data['eonr']
    X_train = data.drop('eonr', axis=1)

    X_train=X_train.values
    y_train=y_train.values

    #Calculate some other hyperparameters based on data.  
    batch_no = len(X_train) // batch_size  #batches
    cols=X_train.shape[1] #Number of columns in input matrix

    #Create the model
    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
    # Assume that we are on a CUDA machine, then this should print a CUDA device:
    print("Executing the model on :",device)

    net = Net(cols)
    #Adam is a specific flavor of gradient decent which is typically better
    optimizer = torch.optim.Adam(net.parameters(), lr=learning_rate)
    #optimizer = torch.optim.SGD(net.parameters(), lr=0.2)
    criterion = torch.nn.MSELoss(size_average=False)  # this is for regression mean squared loss

    running_loss = 0.0
    for epoch in range(num_epochs):
        #Shuffle just mixes up the dataset between epocs
        X_train, y_train = shuffle(X_train, y_train)
        # Mini batch learning
        for i in range(batch_no):
            start = i * batch_size
            end = start + batch_size
            inputs = Variable(torch.FloatTensor(X_train[start:end]))
            labels = Variable(torch.FloatTensor(y_train[start:end]))
            # clear the gradients
            optimizer.zero_grad()
            # compute the model output
            outputs = net(inputs)
            # calculate loss
            loss = criterion(outputs, torch.unsqueeze(labels,dim=1))
            # credit assignment
            loss.backward()
            # update model weights
            optimizer.step()

            # print statistics
            running_loss += loss.item()

        #print('Epoch {}'.format(epoch+1), "loss: ",running_loss)
        running_loss = 0.0
    path = '/home/germanm2/n_policy_box/Data/files_rds/cnn_models/'+ policy + '.pth'
    torch.save(net.state_dict(), path)
    return(net)
        

# from sklearn.preprocessing import MinMaxScaler, StandardScaler
# TrainSet_eonr2_df = pyreadr.read_r("/home/germanm2/n_policy_box/Data/files_rds/TrainSet_eonr2.rds")[None] # also works for RData
# scaler = StandardScaler()
# cols = [c for c in TrainSet_eonr2_df.columns if c not in ['id', 'target', 'wheezy-copper-turtle-magic']]
# TrainSet_eonr2_df[cols] = scaler.fit_transform(TrainSet_eonr2_df[cols])


# from sklearn.preprocessing import MinMaxScaler, StandardScaler
# TrainSet_eonr2_df = pyreadr.read_r("/home/germanm2/n_policy_box/Data/files_rds/TrainSet_eonr2.rds")[None] # also works for RData
# scaler = StandardScaler()
# scaler.fit(TrainSet_eonr2_df)
# TrainSet_eonr_scaled = scaler.transform(TrainSet_eonr2_df)


from sklearn.preprocessing import StandardScaler
TrainSet_eonr2_df = pyreadr.read_r("/home/germanm2/n_policy_box/Data/files_rds/TrainSet_eonr2.rds")[None] # also works for RDatascaler = StandardScaler()
TrainSet_eonr2_df = TrainSet_eonr2_df[high_var+['eonr']]

# First create a copy of your dataframe:
TrainSet_eonr2_scaled = TrainSet_eonr2_df.copy()
# Don't include the Name column in the transformation:
features = TrainSet_eonr2_scaled[high_var]
scaler = StandardScaler().fit(features.values)
features = scaler.transform(features.values)
# Now, don't create a new dataframe but assign the result to those two columns:

TrainSet_eonr2_scaled[high_var] = features
TrainSet_eonr2_scaled

#Build one cnn using the function
# TrainSet_eonr2_df = pyreadr.read_r("/home/germanm2/n_policy_box/Data/files_rds/TrainSet_eonr2.rds")[None] # also works for RData
# TrainSet_eonr2_df = TrainSet_eonr2_df[high_var+['eonr']]
net_return = build_cnn(TrainSet_eonr2_scaled, 'ratio_5')

# #Make predictions with returned model
# prediction_set_aggregated_df = pyreadr.read_r("/home/germanm2/n_policy_box/Data/files_rds/prediction_set_aggregated_dt.rds")[None] # also works for RData

# X_pred = prediction_set_aggregated_df[['rain_30', 'rain_60','rain_90', 't_max_30', 't_max_60', 't_max_90', 't_min_30', 't_min_60', 't_min_90', 'Y_prev',
#  'Y_corn_lt_avg', 'day_sow', 'day_v5', 'lai_v5', 'whc', 'oc_20cm_v5', 'sw_dep_v5', 'n_0_60cm_v5', 'surfaceom_wt_v5', 'sand_40cm', 'clay_40cm']]

# X_pred=X_pred.values

# X = Variable(torch.FloatTensor(X_pred)) 
# y_pred = net_return(X) #This outputs the value for regression
# y_pred=y_pred.data[:,0].numpy()
# y_pred
# prediction_set_aggregated_df['eonr_pred'] = y_pred

# # now let's write a Rds
# pyreadr.write_rds("/home/germanm2/n_policy_box/Data/files_rds/prediction_set_aggregated_cnn_dt.rds", prediction_set_aggregated_df)
# prediction_set_aggregated_df.head()

#Load the saved model
policy = 'ratio_5'
path = '/home/germanm2/n_policy_box/Data/files_rds/cnn_models/'+ policy + '.pth'
net_load = Net(21)
net_load.load_state_dict(torch.load(path))
net_load.eval()

# Load and scale the prediction data
prediction_set_aggregated_df = pyreadr.read_r("/home/germanm2/n_policy_box/Data/files_rds/prediction_set_aggregated_dt.rds")[None] # also works for RData

# First create a copy of your dataframe:
prediction_set_scaled = prediction_set_aggregated_df.copy()
# Don't include the Name column in the transformation:
features = prediction_set_scaled[high_var]
features = scaler.transform(features.values)
# Now, don't create a new dataframe but assign the result to those two columns:

prediction_set_scaled[high_var] = features
X_pred = prediction_set_scaled[high_var]

X_pred=X_pred.values
X = Variable(torch.FloatTensor(X_pred)) 

#Make predictions with the saved model
y_pred = net_load(X) #This outputs the value for regression
y_pred=y_pred.data[:,0].numpy()
y_pred
prediction_set_aggregated_df['eonr_pred'] = y_pred

# now let's write a Rds
pyreadr.write_rds("/home/germanm2/n_policy_box/Data/files_rds/prediction_set_aggregated_cnn_dt.rds", prediction_set_aggregated_df)
prediction_set_aggregated_df.head()

from sklearn.metrics import mean_squared_error

print(prediction_set_aggregated_df['eonr_pred'].min(),
prediction_set_aggregated_df['eonr_pred'].max(),
prediction_set_aggregated_df['eonr_pred'].mean(),
mean_squared_error(prediction_set_aggregated_df['eonr_12'], prediction_set_aggregated_df['eonr_pred']))

# 3228 3152 2966 3292 3184 3192 3254 2993
import matplotlib.pyplot as plt
import numpy as np

x = prediction_set_aggregated_df['eonr_12']
y = prediction_set_aggregated_df['eonr_pred']

plt.scatter(x, y, c="g", alpha=0.5, marker=r'$\clubsuit$',
            label="Luck")
plt.legend(loc='upper left')
plt.show()

# Make a function that loads the saved model and does predictions

prediction_set_aggregated_df = pyreadr.read_r("/home/germanm2/n_policy_box/Data/files_rds/prediction_set_aggregated_dt.rds")[None]


def predict_cnn(prediction_set_aggregated_df, policy):
    #Load the saved model
    #policy = 'ratio_5'
    path = '/home/germanm2/n_policy_box/Data/files_rds/cnn_models/'+ policy + '.pth'
    net_load = Net(21)
    net_load.load_state_dict(torch.load(path))
    net_load.eval()
    net_load.state_dict()
    X_pred = prediction_set_aggregated_df[high_var]
    X_pred=X_pred.values
    X_pred
    X = Variable(torch.FloatTensor(X_pred)) 
    y_pred = net_load(X) #This outputs the value for regression
    y_pred=y_pred.data[:,0].numpy()
    y_pred
    prediction_set_aggregated_df['eonr_pred'] = y_pred
    return(prediction_set_aggregated_df)


# Use the function
prediction_set_aggregated_df2 = predict_cnn(prediction_set_aggregated_df, 'ratio_5')
prediction_set_aggregated_df2
